refactor: 修改BaseApiController位置

+++ src/TTMS.DTO/Request/UserRequest.cs
@@ -8,10 +8,10 @@
         /// <summary>
         /// 用户id
         /// </summary>
-        public long? Id { get; set; }
+        public int? Id { get; set; }
 
         /// <summary>
-        /// 账号
+        /// 账户名
         /// </summary>
         public string? Account { get; set; }
 
@@ -31,14 +31,14 @@
         public RoleType? RoleId { get; set; }
 
         /// <summary>
-        /// 账号状态
+        /// 账户状态
         /// </summary>
         public bool? State { get; set; }
 
         /// <summary>
         /// 创建人
         /// </summary>
-        public long? CreateBy { get; set; }
+        public int? CreateBy { get; set; }
 
         /// <summary>
         /// 创建时间
@@ -51,17 +51,17 @@
         public DateTime? CreateTimeEnd { get; set; }
 
         /// <summary>
-        /// 更新人
+        /// 最后更新人
         /// </summary>
-        public long? UpdateBy { get; set; }
+        public int? UpdateBy { get; set; }
 
         /// <summary>
-        /// 更新时间
+        /// 最后更新时间
         /// </summary>
         public DateTime? UpdateTimeStart { get; set; }
 
         /// <summary>
-        /// 更新时间
+        /// 最后更新时间
         /// </summary>
         public DateTime? UpdateTimeEnd { get; set; }
     }
@@ -72,7 +72,7 @@
     public class CreateUserRequest
     {
         /// <summary>
-        /// 账号
+        /// 账户
         /// </summary>
         [Required(ErrorMessage = "Account is required.")]
         public string? Account { get; set; }
@@ -101,7 +101,7 @@
         public RoleType RoleId { get; set; }
 
         /// <summary>
-        /// 账号状态
+        /// 账户状态
         /// </summary>
         [Required(ErrorMessage = "State is required.")]
         public bool State { get; set; }
@@ -116,7 +116,7 @@
         /// 用户Id
         /// </summary>
         [Required(ErrorMessage = "Id is required.")]
-        public long? Id { get; set; }
+        public int? Id { get; set; }
     }
 
     /// <summary>
