+++ src/TTMS.DTO/Response/UserResponse.cs
@@ -8,10 +8,10 @@
         /// <summary>
         /// 用户id
         /// </summary>
-        public long Id { get; set; }
+        public int Id { get; set; }
 
         /// <summary>
-        /// 账号
+        /// 账户名
         /// </summary>
         public string? Account { get; set; }
 
@@ -31,14 +31,14 @@
         public RoleType RoleId { get; set; }
 
         /// <summary>
-        /// 账号状态
+        /// 账户状态
         /// </summary>
         public bool State { get; set; }
 
         /// <summary>
         /// 创建人
         /// </summary>
-        public long CreateBy { get; set; }
+        public int CreateBy { get; set; }
 
         /// <summary>
         /// 创建时间
@@ -46,12 +46,12 @@
         public DateTime CreateTime { get; set; }
 
         /// <summary>
-        /// 更新人
+        /// 最后更新人
         /// </summary>
-        public long UpdateBy { get; set; }
+        public int UpdateBy { get; set; }
 
         /// <summary>
-        /// 更新时间
+        /// 最后更新时间
         /// </summary>
         public DateTime UpdateTime { get; set; }
     }
